requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
langchain-community
langgraph
python-dotenv
requests
beautifulsoup4
yfinance
pandas
tavily-python